This is a mac app
This is the python script compiled to a mac app that you can unzip and just use. Usually open source projects like this require you to compile the script yourself but I understand that some people who could benefit from this may not want to do that. So I'm providing it here.